+++ spec/time_spec.rb
@@ -296,5 +296,19 @@ describe EtOrbi::EoTime do
       expect((ot0 - ot1).to_i).to eq(-222)
     end
   end
+
+  describe '#wday_in_month' do
+
+    it 'computes the wday in month interval' do
+
+      t = EtOrbi::EoTime.new(1193898300, 'Europe/Paris')
+
+      expect(t.wday_in_month).to eq([ 1, -5 ])
+
+      t = EtOrbi.make_time('2017-06-21 Europe/Paris')
+
+      expect(t.wday_in_month).to eq([ 3, -2 ])
+    end
+  end
 end